Moved Price to a standalone file.
